django/requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.21.0
isort==5.10.1
django-split-settings==1.2.0
psycopg2==2.9.5
gunicorn==20.1.0
djangorestframework==3.14.0